feat: refactor + mock is broken
