+++ system-contract-testing/package.json
@@ -9,8 +9,10 @@
   "author": "Hedera Smart Contracts Team",
   "license": "Apache-2.0",
   "dependencies": {
+    "@hashgraph/proto": "^2.22.0",
     "@nomicfoundation/solidity-analyzer": "^0.1.2",
     "bn.js": "^5.2.2",
+    "chai": "^4.5.0",
     "dotenv": "^17.2.0",
     "elliptic": "^6.6.1",
     "mcl-wasm": "^2.0.0",
