+++ system-contract-testing/test/hip-1215/utils/hip1215-mock.js
